refactor(cli): extract analyzer scoring helpers
